# api.yaml
openapi: 3.0.3
info:
  title: Avito Kitchen API
  description: HTTP API backend-сервиса площадки «Авито.Кухня» для клиентов и партнерских ресторанов.
  version: 1.0.0

servers:
  - url: http://localhost:8000
    description: Локальный сервер разработки

paths:
  /restaurants:
    get:
      summary: Получение каталога заведений
      description: Возвращает список всех зарегистрированных заведений, отсортированных по названию в алфавитном порядке.
      operationId: getRestaurants
      responses:
        "200":
          description: Успешный возврат списка ресторанов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Restaurant"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /restaurants/{id}/menu:
    get:
      summary: Получение меню ресторана
      description: Возвращает список активных и доступных для заказа позиций блюд выбранного заведения.
      operationId: getRestaurantMenu
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор заведения (UUID)
          schema:
            type: string
            format: uuid
            example: 00000000-0000-0000-0000-000000000001
      responses:
        "200":
          description: Успешный возврат списка доступных блюд
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MenuItem"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /orders:
    post:
      summary: Оформление нового заказа
      description: Создает заказ пользователя с валидацией доступности всех позиций и принадлежности одному ресторану.
      operationId: createOrder
      parameters:
        - name: X-User-ID
          in: header
          required: true
          description: Идентификатор пользователя (UUID)
          schema:
            type: string
            format: uuid
            example: aaaaaaaa-aaaa-aaaa-aaaa-aaaaaaaaaaaa
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Заказ успешно создан и переведен в статус Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /orders/{id}:
    get:
      summary: Получение информации о заказе
      description: Возвращает информацию о заказе, текущем статусе и позициях по его ID.
      operationId: getOrderByID
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор заказа (UUID)
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Информация о заказе найдена
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "404":
          $ref: "#/components/responses/NotFoundError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /restaurant/orders:
    get:
      summary: Получение очереди заказов ресторана
      description: Возвращает список заказов, закрепленных за рестораном, с опциональной фильтрацией по статусу.
      operationId: getRestaurantOrders
      parameters:
        - name: X-Restaurant-ID
          in: header
          required: true
          description: Идентификатор ресторана (UUID)
          schema:
            type: string
            format: uuid
            example: 00000000-0000-0000-0000-000000000001
        - name: status_id
          in: query
          required: false
          description: Фильтр по числовому идентификатору статуса (1 - Created, 2 - Accepted, 3 - Cooking, 4 - Ready, 5 - Cancelled)
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: Список заказов ресторана
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

  /restaurant/orders/{id}/status:
    patch:
      summary: Обновление статуса заказа кухней
      description: Переводит заказ на следующий этап жизненного цикла с проверкой матрицы переходов.
      operationId: updateOrderStatus
      parameters:
        - name: id
          in: path
          required: true
          description: Идентификатор заказа (UUID)
          schema:
            type: string
            format: uuid
        - name: X-Restaurant-ID
          in: header
          required: true
          description: Идентификатор ресторана (UUID)
          schema:
            type: string
            format: uuid
            example: 00000000-0000-0000-0000-000000000001
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrderStatusRequest"
      responses:
        "200":
          description: Статус заказа успешно обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: updated
        "400":
          $ref: "#/components/responses/BadRequestError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          $ref: "#/components/responses/InternalServerError"

components:
  schemas:
    Restaurant:
      type: object
      required:
        - id
        - name
        - created_at
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Пиццерия у Тома
        description:
          type: string
          example: Готовим настоящую итальянскую пиццу в дровяной печи.
        created_at:
          type: string
          format: date-time

    MenuItem:
      type: object
      required:
        - id
        - restaurant_id
        - name
        - price
        - is_available
      properties:
        id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        name:
          type: string
          example: Пицца Пепперони
        price:
          type: integer
          description: Стоимость в копейках
          example: 55000
        is_available:
          type: boolean
          example: true

    OrderItemInput:
      type: object
      required:
        - menu_item_id
        - quantity
      properties:
        menu_item_id:
          type: string
          format: uuid
          example: 11111111-1111-1111-1111-111111111101
        quantity:
          type: integer
          minimum: 1
          example: 2

    CreateOrderRequest:
      type: object
      required:
        - restaurant_id
        - items
      properties:
        restaurant_id:
          type: string
          format: uuid
          example: 00000000-0000-0000-0000-000000000001
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItemInput"

    OrderItem:
      type: object
      required:
        - id
        - order_id
        - menu_item_id
        - quantity
        - price_at_purchase
      properties:
        id:
          type: string
          format: uuid
        order_id:
          type: string
          format: uuid
        menu_item_id:
          type: string
          format: uuid
        quantity:
          type: integer
          example: 2
        price_at_purchase:
          type: integer
          description: Фиксированная цена блюда в копейках на момент покупки
          example: 55000

    Order:
      type: object
      required:
        - id
        - user_id
        - restaurant_id
        - status_id
        - items
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        restaurant_id:
          type: string
          format: uuid
        status_id:
          type: integer
          description: "Идентификатор статуса: 1 (Created), 2 (Accepted), 3 (Cooking), 4 (Ready), 5 (Cancelled)"
          example: 1
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"

    UpdateOrderStatusRequest:
      type: object
      required:
        - status_id
      properties:
        status_id:
          type: integer
          description: Целевой статус заказа (2 - Accepted, 3 - Cooking, 4 - Ready, 5 - Cancelled)
          example: 2

    ErrorResponse:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          example: Error occurred

  responses:
    BadRequestError:
      description: Некорректные параметры запроса, невалидный UUID, тело JSON или нарушение перехода статусов
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    UnauthorizedError:
      description: Отсутствует или некорректен заголовок авторизации (X-User-ID / X-Restaurant-ID)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    NotFoundError:
      description: Запрашиваемый ресурс (заказ или ресторан) не найден
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"

    InternalServerError:
      description: Внутренняя ошибка сервера при обращении к базе данных
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"